feat(router): introduce providePlatformNavigationRouter and separate HistoryStateManager
Separate router state managers into distinct provider functions:
- Introduce providePlatformNavigationRouter for Navigation API support.
- Move HistoryStateManager to its own file so applications using
  providePlatformNavigationRouter avoid paying the bundle size cost of
  HistoryStateManager and History API.
- Remove experimental withExperimentalPlatformNavigation() feature.
