+++ packages/router/src/index.ts
@@ -74,6 +74,7 @@ export {ViewTransitionInfo, ViewTransitionsFeatureOptions} from './utils/view_tr
 export * from './models_deprecated';
 export {Navigation, NavigationExtras, UrlCreationOptions} from './navigation_transition';
 export {DefaultTitleStrategy, TitleStrategy} from './page_title_strategy';
+export {providePlatformNavigationRouter} from './provide_platform_navigation_router';
 export {
   AutoCleanupInjectorsFeature,
   ComponentInputBindingFeature,
@@ -94,9 +95,8 @@ export {
   withDebugTracing,
   withDisabledInitialNavigation,
   withEnabledBlockingInitialNavigation,
-  withExperimentalAutoCleanupInjectors,
   withAutoCleanupInjectors,
-  withExperimentalPlatformNavigation,
+  withExperimentalAutoCleanupInjectors,
   withHashLocation,
   withInMemoryScrolling,
   withNavigationErrorHandler,
