+++ packages/router/test/integration/integration.spec.ts
@@ -35,7 +35,8 @@ import {
   RoutesRecognized,
 } from '../../index';
 
-import {provideRouter, withExperimentalPlatformNavigation} from '../../src/provide_router';
+import {providePlatformNavigationRouter} from '../../src/provide_platform_navigation_router';
+import {provideRouter} from '../../src/provide_router';
 import {
   BlankCmp,
   CollectParamsCmp,
@@ -85,12 +86,9 @@ for (const browserAPI of ['navigation', 'history'] as const) {
         imports: [...ROUTER_DIRECTIVES, TestModule],
         providers: [
           {provide: Console, useValue: noopConsole},
-          provideRouter(
-            [{path: 'simple', component: SimpleCmp}],
-            browserAPI === 'navigation'
-              ? withExperimentalPlatformNavigation()
-              : (makeEnvironmentProviders([]) as any),
-          ),
+          browserAPI === 'navigation'
+            ? providePlatformNavigationRouter([{path: 'simple', component: SimpleCmp}])
+            : provideRouter([{path: 'simple', component: SimpleCmp}]),
         ],
       });
     });
@@ -964,16 +962,16 @@ for (const browserAPI of ['navigation', 'history'] as const) {
     });
 
     navigationErrorsIntegrationSuite(browserAPI);
-    eagerUrlUpdateStrategyIntegrationSuite();
+    eagerUrlUpdateStrategyIntegrationSuite(browserAPI);
     duplicateInFlightNavigationsIntegrationSuite(browserAPI);
     navigationIntegrationTestSuite(browserAPI);
     routeDataIntegrationSuite();
     routerLinkIntegrationSpec();
     redirectsIntegrationSuite(browserAPI);
-    guardsIntegrationSuite();
+    guardsIntegrationSuite(browserAPI);
     routerEventsIntegrationSuite();
-    routerLinkActiveIntegrationSuite();
+    routerLinkActiveIntegrationSuite(browserAPI);
     lazyLoadingIntegrationSuite(browserAPI);
-    routeReuseIntegrationSuite();
+    routeReuseIntegrationSuite(browserAPI);
   });
 }
